Added valobject.pas as need by spot class.
Continuing work on grafting in new RB/PSKR code.  Starting to come together, but there's still a lot of old hooks to rbc.pas that need to be weeded out.  Historical not.  rbc.pas Is some of the oldest code in the tree. It predates JT65-HF and came from rbc.exe as used with WSJT.  As much as it had to go away it's still sad to see its end.